fix: move wallet connect context files closer together
