+++ packages/apps/human-app/frontend/src/modules/operator/views/sign-up/connect-wallet.page.tsx
@@ -2,7 +2,7 @@ import { Grid, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { Navigate } from 'react-router-dom';
 import { PageCard } from '@/shared/components/ui/page-card';
-import { useWalletConnect } from '@/shared/hooks/use-wallet-connect';
+import { useWalletConnect } from '@/shared/contexts/wallet-connect';
 import { Alert } from '@/shared/components/ui/alert';
 import { getErrorMessageForError } from '@/shared/errors';
 import { Button } from '@/shared/components/ui/button';
