+++ packages/apps/human-app/frontend/src/shared/contexts/wallet-connect/index.ts
@@ -1,0 +1,3 @@
+export * from './wallet-connect';
+export * from './use-wallet-connect';
+export * from './require-wallet-connect';
