+++ packages/apps/human-app/frontend/src/shared/contexts/wallet-connect/require-wallet-connect.tsx
@@ -1,9 +1,9 @@
 import { useLocation, Navigate } from 'react-router-dom';
 import { createContext } from 'react';
-import { useWalletConnect } from '@/shared/hooks/use-wallet-connect';
 import { routerPaths } from '@/router/router-paths';
 import { PageCardLoader } from '@/shared/components/ui/page-card';
-import { type WalletConnectContextConnectedAccount } from '@/shared/types/wallet-connect';
+import { type WalletConnectContextConnectedAccount } from './types';
+import { useWalletConnect } from './use-wallet-connect';
 
 export const AuthWeb3Context =
   createContext<WalletConnectContextConnectedAccount | null>(null);
