+++ packages/apps/human-app/frontend/src/shared/contexts/wallet-connect/types.ts
