+++ packages/apps/human-app/frontend/src/shared/contexts/wallet-connect/wallet-connect.tsx
@@ -21,7 +21,7 @@ import { JsonRpcError } from '@/modules/smart-contracts/json-rpc-error';
 import {
   type WalletConnectContextConnectedAccount,
   type WalletConnectContextDisconnectedAccount,
-} from '../types/wallet-connect';
+} from './types';
 
 const projectId = env.VITE_WALLET_CONNECT_PROJECT_ID;
